rename DaoFactory to ContractUtil
