+++ test/adapters/configuration.test.js
@@ -24,7 +24,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
  */
-const { sha3, toBN } = require("../../utils/DaoFactory.js");
+const { sha3, toBN } = require("../../utils/ContractUtil.js");
 
 const {
   deployDefaultDao,
