+++ test/adapters/coupon-onboarding.test.js
@@ -30,7 +30,7 @@ const {
   SHARES,
   GUILD,
   ETH_TOKEN,
-} = require("../../utils/DaoFactory.js");
+} = require("../../utils/ContractUtil.js");
 
 const {
   deployDefaultDao,
