+++ test/adapters/offchain.voting.test.js
@@ -30,7 +30,7 @@ const {
   sharePrice,
   remaining,
   SHARES,
-} = require("../../utils/DaoFactory.js");
+} = require("../../utils/ContractUtil.js");
 
 const {
   proposalIdGenerator,
