+++ test/adapters/tribute-nft.test.js
@@ -31,7 +31,7 @@ const {
   GUILD,
   ETH_TOKEN,
   sha3,
-} = require("../../utils/DaoFactory.js");
+} = require("../../utils/ContractUtil.js");
 
 const {
   deployDefaultNFTDao,
