+++ test/extensions/bank.test.js
@@ -24,7 +24,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
  */
-const { ETH_TOKEN, sha3 } = require("../../utils/DaoFactory.js");
+const { ETH_TOKEN, sha3 } = require("../../utils/ContractUtil.js");
 
 const {
   deployDefaultDao,
