+++ test/helpers/fair-share-helper.test.js
@@ -20,7 +20,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
  */
-const { toWei } = require("../../utils/DaoFactory.js");
+const { toWei } = require("../../utils/ContractUtil.js");
 
 const {
   TestFairShareCalc,
