+++ utils/TestUtils.js
@@ -1,7 +1,7 @@
 // Whole-script strict mode syntax
 "use strict";
 
-const { fromUtf8, toBN, SHARES, LOOT } = require("./DaoFactory.js");
+const { fromUtf8, toBN, SHARES, LOOT } = require("./ContractUtil.js");
 
 const { expect, advanceTime } = require("./OZTestUtil.js");
