+++ utils/TruffleUtil.js
@@ -25,7 +25,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
  */
 
-const { contracts } = require("./DaoFactory");
+const { contracts } = require("./ContractUtil");
 
 const truffleContracts = getContracts();
